+++ modules/integration/test-results-service/src/main/java/org/wso2/sp/events/Event.java
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2016, WSO2 Inc. (http://wso2.com) All Rights Reserved.
+ * Copyright (c) 2017, WSO2 Inc. (http://wso2.com) All Rights Reserved.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.wso2.sp.event;
+package org.wso2.sp.events;
 
 import org.codehaus.jackson.map.annotate.JsonRootName;
 import java.io.Serializable;
